// src/app/models/usuarios.ts
export class Usuarios {
    id?: string;
    Apellido: string;
    Correo: string;
    Foto2: string;
    FotoPerfil: string;
    Matricula: string;
    Nombre: string;
    Premium: boolean;
    Publicaciones: number;
    Rol: string;
    Telefono: string;
    Universidad: string;
    Verificacion: boolean;
    Estado: string;
    EsperaAyudantia: boolean;
    Anuncio: boolean;
    EsperaPremium:boolean;
    Cursos:number;
    AyudantiaAceptada:boolean;
    Citas:boolean;

    Descripcion: string;
    Busca: string;
    Sexo: string;
    Nacimiento:string;
}